+++ libft/ft_strlcpy.c
@@ -2,8 +2,8 @@
 
 size_t  ft_strlcpy(char *dest, const char *src, size_t size)
 {
-    size_t len;
-    int i;
+    size_t  len;
+    size_t  i;
 
     i = 0;
     if (!src || !dest)
